requirements.txt
fastapi
uvicorn[standard]
sqlmodel>=0.0.16
passlib[bcrypt]
python-jose[cryptography]
